requirements.txt
agora-token-builder==1.0.0
asgiref==3.7.2
beautifulsoup4==4.12.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap5==0.7
cryptography==41.0.7
cycler==0.12.1
Django==4.2.7
django-admin-interface==0.19.1
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-colorfield==0.7.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-dirtyfields==1.9.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-smart-selects==1.6.0
django-tables2==2.4.1
djangorestframework==3.14.0
filelock==3.13.1
firebase-admin==6.3.0
fonttools==4.46.0
fsspec==2023.12.2
gcloud==0.18.3
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
jws==0.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
mysqlclient==2.2.0
networkx==3.2.1
numpy==1.26.2
oauth2client==4.1.3
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==9.3.0
proto-plus==1.22.3
protobuf==4.25.1
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pynput==1.7.6
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jwt==2.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.227
uritemplate==4.1.1
urllib3==1.26.18
zipp==3.17.0
